+++ tests/testthat/test-all.R
@@ -1,11 +1,17 @@
 
 context("Get")
 
-test_that("search_eurostat finds",{
+test_that("get_eurostat includes time and value",{
   skip_on_cran()
-  expect_true(all(c("time", "values") %in% names(get_eurostat("namq_aux_lp", cache = FALSE))))
+  expect_true(all(c("time", "values") %in% 
+                    names(get_eurostat("namq_aux_lp", cache = FALSE))))
 })
 
+test_that("get_eurostat works with multi-frequency",{
+  skip_on_cran()
+  expect_error(get_eurostat("avia_gonc", cache = FALSE))
+  matches("-01-01", unique(get_eurostat("avia_gonc", select_time = "Y" , cache = FALSE)$time))
+})
 
 context("Search")